LPS-55584 Move TunnelingServletAuthVerifier into OSGi module
