+++ portal-impl/src/portal.properties
@@ -3663,8 +3663,8 @@
     #
     # TunnelingServletAuthVerifier
     #
-    auth.verifier.TunnelingServletAuthVerifier.hosts.allowed=255.255.255.255
-    auth.verifier.TunnelingServletAuthVerifier.urls.includes=/api/liferay/do
+    #auth.verifier.TunnelingServletAuthVerifier.hosts.allowed=255.255.255.255
+    #auth.verifier.TunnelingServletAuthVerifier.urls.includes=/api/liferay/do
 
 ##
 ## Authentication Token
